+++ examples/compute/instance/front_worker/base-image.tf
@@ -1,1 +1,0 @@
-../../../images/image/datasource/main.tf