+++ examples/lb/members/base-computes.tf
@@ -1,1 +1,0 @@
-../../compute/instance/front_worker/main.tf